+++ include/visual_servo/Servo.h
@@ -22,13 +22,13 @@ class Servo
 private:
     tf2_ros::Buffer tfBuffer_;
     tf2_ros::TransformListener *tfListener_;
-
+    Destination_t EndDestinationDelta;
 public:
     Servo();
     virtual ~Servo();
 
-    void getTransform(const std::string source_frame,const std::string target_frame,Eigen::Affine3d &transformMatrix);
-    Eigen::Affine3d getTransform(const std::string source_frame,const std::string target_frame);
+    void getTransform(const std::string & source_frame,const std::string & target_frame,Eigen::Affine3d &transformMatrix);
+    Eigen::Affine3d getTransform(const std::string & source_frame,const std::string & target_frame);
     /*
      * Function computes the end effector motion when achieving camera motion as a transform matrix
      * @param Trans_C2T         [the target pose with respect to camera described as a transform matrix]
@@ -36,7 +36,7 @@ public:
      * @param ExpectTrans_C2T   [the desired target pose with respect to camera described as a transform matrix]
      * @return the end effector motion described as a homogeneous matrix and its position error
      */
-    Destination_t getCameraEE (Eigen::Affine3d Trans_C2T,Eigen::Affine3d Trans_E2C, Eigen::Affine3d ExpectTrans_C2T,double lambda=0.8);
+    const Destination_t & getCameraEE (Eigen::Affine3d Trans_C2T,Eigen::Affine3d Trans_E2C, Eigen::Affine3d ExpectTrans_C2T,double lambda=0.8);
 
 };