+++ Change Log.md
@@ -13,3 +13,11 @@
     	1. 添加了一个在后台工作的参数服务器监视器并且给出了两种获取参数值的封装方法
      	2. 添加了机械臂软件急停功能
      	3. 更改了新的模型
+- **Header**: fcfb478
+  - **comments** : fix go charge bugs and give a safe way to get parameters from listener
+  - **Actual Change**:
+    	1. 修复了充电命令
+    	2. 给出了获取参数值的安全方法
+     	3. 注意：没有上传parameter.yaml
+
+